% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombcVizhelper-package.R
\docType{package}
\name{ancombcVizhelper-package}
\alias{ancombcVizhelper}
\alias{ancombcVizhelper-package}
\title{ancombcVizhelper: Tidy and Visualize Results from ANCOM-BC2}
\description{
Provides tools to tidy, summarize, and visualize differential abundance results generated by ANCOM-BC2. The package converts ANCOM-BC2 model output into a standardized tidy format and generates customizable, publication-ready visualizations, including log-fold change plots, heatmaps, and comparison-specific summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KitHubb/ancombcVizhelper}
  \item Report bugs at \url{https://github.com/KitHubb/ancombcVizhelper/issues}
}

}
\author{
\strong{Maintainer}: So-Yeon Kim \email{kim.soyeon.bio@gmail.com}

Authors:
\itemize{
  \item So-Yeon Kim \email{kim.soyeon.bio@gmail.com}
}

}
\keyword{internal}
